added seperate component for questions of topic
