+++ frontend/src/App.jsx
@@ -6,13 +6,13 @@ import "./App.css";
 import LandingPage from "./pages/LandingPage.jsx";
 import LoginPage from "./pages/LoginPage.jsx";
 import SignupPage from "./pages/SignupPage.jsx";
-import DashboardPage from "./pages/DashboardPage.jsx";
 import ContentPage from "./pages/ContentPage.jsx";
 import ProblemPage from "./pages/ProblemPage.jsx";
 import ErrorPage from "./pages/ErrorPage.tsx";
 import RootPage from "./pages/RootPage.jsx";
 import HomeRootPage from "./pages/HomeRootPage.jsx";
-import { loader as topicsLoader } from "./pages/DashboardPage.jsx";
+import DashboardPage from "./pages/DashboardPage.tsx";
+import { loader as topicsLoader } from "./pages/DashboardPage.tsx";
 import { loader as problemLoader } from "./pages/ProblemPage.jsx";
 import { loader as contentLoader } from "./pages/ContentPage.jsx";
 import { action as signupAction } from "./pages/SignupPage.jsx";
